HWKAPM-518 Refactor some model objects and cache APIs into common modules as they are required by additional implementations
