+++ processors/tracecompletiontime-deriver/src/test/java/org/hawkular/apm/processor/tracecompletiontime/TestCommunicationDetailsCache.java
@@ -22,6 +22,7 @@ import java.util.List;
 import java.util.Map;
 
 import org.hawkular.apm.api.model.events.CommunicationDetails;
+import org.hawkular.apm.server.api.services.CommunicationDetailsCache;
 
 /**
  * @author gbrown
