+++ server/kafka/src/main/java/org/hawkular/apm/server/kafka/CommunicationDetailsCacheKafka.java
@@ -22,8 +22,8 @@ import java.util.logging.Logger;
 
 import org.hawkular.apm.api.model.events.CommunicationDetails;
 import org.hawkular.apm.api.services.ServiceResolver;
-import org.hawkular.apm.processor.tracecompletiontime.CommunicationDetailsCache;
 import org.hawkular.apm.server.api.services.CacheException;
+import org.hawkular.apm.server.api.services.CommunicationDetailsCache;
 import org.hawkular.apm.server.api.task.AbstractProcessor;
 import org.hawkular.apm.server.api.task.Processor.ProcessorType;
 import org.hawkular.apm.server.api.task.RetryAttemptException;
