+++ server/kafka/src/main/java/org/hawkular/apm/server/kafka/CommunicationDetailsDeriverKafka.java
@@ -22,8 +22,8 @@ import org.hawkular.apm.api.model.events.CommunicationDetails;
 import org.hawkular.apm.api.model.trace.Trace;
 import org.hawkular.apm.api.services.ServiceResolver;
 import org.hawkular.apm.processor.communicationdetails.CommunicationDetailsDeriver;
-import org.hawkular.apm.processor.communicationdetails.ProducerInfoCache;
 import org.hawkular.apm.server.api.services.CommunicationDetailsPublisher;
+import org.hawkular.apm.server.api.services.ProducerInfoCache;
 
 import com.fasterxml.jackson.core.type.TypeReference;
